# Add APIs to offload states of model, optimizer, and engine