--- a/deepspeed/runtime/utils.py
+++ b/deepspeed/runtime/utils.py
@@ -1067,18 +1067,21 @@ def get_norm_with_moe_layers(non_expert_norm, mpu, expert_tensors, norm_type=2):
     return total_norm
 
 
+def _make_offload_state_key(key):
+    return f"{key}_offload_buffer"
+
+
 def offload_adam_states(optimizer, device, pin_memory: bool = False, non_blocking: bool = False):
     """Move optimizer states to device. Note that this assumes the state structure of DeepSpeed Adam."""
 
     def move_key(state, key):
-        if pin_memory:
-            pin_mem_key = f"{key}_pin_memory"
-            if pin_mem_key not in state:
-                state[pin_mem_key] = get_accelerator().pin_memory(torch.empty_like(state[key], device=device))
-            state[pin_mem_key].copy_(state[key], non_blocking=non_blocking)
-            state[key].data = state[pin_mem_key]
-        else:
-            state[key].data = state[key].to(device, non_blocking=non_blocking)
+        offload_buf_key = _make_offload_state_key(key)
+        if offload_buf_key not in state:
+            state[offload_buf_key] = torch.empty_like(state[key], device=device)
+            if pin_memory:
+                state[offload_buf_key] = get_accelerator().pin_memory(state[offload_buf_key])
+        state[offload_buf_key].copy_(state[key], non_blocking=non_blocking)
+        state[key].data = state[offload_buf_key]
 
     for _, state in optimizer.state.items():
         if "exp_avg" in state:
@@ -1087,13 +1090,11 @@ def offload_adam_states(optimizer, device, pin_memory: bool = False, non_blockin
             move_key(state, "exp_avg_sq")
 
 
-def offload_adam_states_back(optimizer, device, non_blocking: bool = False):
+def reload_adam_states(optimizer, device, non_blocking: bool = False):
     """Move optimizer states to device. Note that this assumes the state structure of DeepSpeed Adam."""
 
     def move_back_key(state, key):
-        pin_mem_key = f"{key}_pin_memory"
-        src_key = pin_mem_key if pin_mem_key in state else key
-        state[key].data = state[src_key].to(device, non_blocking=non_blocking)
+        state[key].data = state[_make_offload_state_key(key)].to(device, non_blocking=non_blocking)
 
     for _, state in optimizer.state.items():
         if "exp_avg" in state: