@tjruwase 2024-09-04
@tohtana, need to add offload APIs to readthedocs, similar to 
https://deepspeed.readthedocs.io/en/latest/zero3.html#gpu-memory-management