@tohtana 2024-09-12
Thank you for reporting, @kfertakis!

I have an [example script](https://github.com/microsoft/DeepSpeedExamples/tree/tohtana/offload_states/training/offload_states) showing the usage of the APIs. Can you try this?
I suspect that ZeRO3 fails to clean the partitioning status for some models. I would like to clarify that your issue is model specific or not.