@tohtana 2024-09-17
Hi @kfertakis, I tried this [example](https://github.com/microsoft/DeepSpeedExamples/tree/tohtana/offload_states/training/offload_states) with a 4B model but it worked. Can you try this on your environment?
It would be also great if you could offer us a simple repo.