@tjruwase 2024-09-18
> `in flight` exception when trying to offload the model again. I thus wanted to ask whether you think this has something to do with a misuse of the API from my end or if you could provide some further context. I'm providing the relevant stack trace below: Thank you again.

@tohtana, I wonder if it is useful to expose `validate_device()` functionality as a deepspeed utility, so that clients can check/confirm the offload status at arbitrary points in their code? 
https://github.com/microsoft/DeepSpeed/blob/8f81634e45a4f28dfa1abe6ee9c6eec5ac184618/tests/unit/runtime/zero/test_offload_states.py#L20

Similar to how `see_memory_usage` enables inspection of HBM/DRAM usage at any point, we could provide mechanisms for offload status. Perhaps we need something like `see_offload_status` that displays the mapping of params, grads, and optimizer to {HBM, DRAM, NVMe}. 

@kfertakis, would love to get your thoughts as well on whether any of the above would be useful? Thanks!