@kfertakis 2024-09-18
Hey, thanks for the comments.

@tohtana, I've tried the example you provided and it does seem to work so I'm sharing a [fork](https://github.com/kfertakis/DeepSpeedExamples/tree/kfertakis/test_offload_states) of the DeepSpeed-Examples repo to showcase the problem. I've modified the DeepSpeed-Chat code to use `offload_states`. After you prepare an environment with the right deepspeed version for the new API and also install DeepSpeed-Chat, you can then run the following:

```
deepspeed --num_gpus=1 ./applications/DeepSpeed-Chat/training/step3_rlhf_finetuning/main.py --actor_model_name_or_path facebook/opt-1.3b --critic_model_name_or_path facebook/opt-1.3b --actor_zero_stage 3 --critic_zero_stage 3 --num_padding_at_beginning 1 --data_path Dahoas/rm-static --per_device_generation_batch_size 2 --per_device_training_batch_size 2 --generation_batches 1 --ppo_epochs 1 --max_answer_seq_len 512 --max_prompt_seq_len 512 --gradient_accumulation_steps 1 --actor_dropout 0.0 --deepspeed --dtype bf16 --enable_hybrid_engine --offload_test
```
this should lead to the `RuntimeError: param {} still in flight` that I mentioned. Any thoughts on this would be much appreciated. 

@tjruwase thanks for the reference. Current problem aside, I can see how the helper functions can be useful in the future for ensuring consistency. thanks.