@tohtana 2024-09-21
Hi @kfertakis, thank you for sharing the repro. It seems that the actual issue is related to ZeRO3's prefetching.

I opened #6557 as a workaround to address this issue. Can you try the branch `tohtana/clean_up_prefetch_param`? It also includes the offloading APIs. You can just switch to it.