@kfertakis 2024-09-25
I also wanted to ask if the offloading functionality could be extended to support `DeepSpeedCPUAdam` optimiser, besides `FusedAdam`, in the future for offloading a model with an optimizer which is already offloaded to the CPU? Thank you